feat(SwiftUI Account Picker): WACO-38 Continue development of Account Picker UI and view models (#2555)
In a first and previous iteration, a very basic version of the Account Picker was created with TCA / SwiftUI. In this second iteration, we have tried to detail the UI, but mainly to create the base of the model that should be an input for the information that is required to recreate the view. For example, the type of Header that the Picker must show or the different types of rows. It is important to note that the final form of the model will depend on the final integration with Ribs. But with this PR it is intended to present an advance or a possible proposal of what the new model would look like.