chore(SPM): AuthenticationFeature to 📦 Swift Package Manager  (#2562)
* chore(SPM): AuthenticationFeature

* sh scripts/format-code.sh

* Remove rogue file

* rename AuthenticationFeature -> FeatureAuthentication

* Point RxSwift to ReactiveX/RxSwift

We have seen a few issues when resolving RxSwift packages, this may be
down to the nature of multiple sources of RxSwift being defined. This
is just to resolve the discrepancy between the sources and may not
resolve the issue we have been seeing. The issue may be a SwiftPM bug:
https://bugs.swift.org/browse/SR-12303

* Update Package.resolved

* sh scripts/format-code.sh

* Add a runtime check for BuildFlag.isInternal

* sh scripts/format-code.sh