feat(Account Picker): WACO-49 TCA/SwiftUI Account Picker Rows UI (#2574)
The initial UI model contemplated only one type of row. With this development, the other 3 models available in the original Picker are added. The existing row has also been modified to accommodate the current development. Each Row has its own model and a particular view. Importantly, models can continue to vary as development continues, and views can also be improved / refined in more detail. This PR is not intended to be a final result due to the degree of advancement that the Account Picker currently has.