feat(SwiftUI New Onboarding Tour): WACO-52 Creation of base 4 screen carousel system (#2594)
To start the new onboarding tour, a new Swift package (FeatureTour) has been created. This will encapsulate everything necessary for the tour, written in TCA / SwiftUI.
At the moment, a basic view has been developed that contains a carousel with its four base views. Each of the views will be configured later, now they only have a different color. In the next developments, each view of the carousel will be detailed, as well as all the actions and effects associated with them.