feat(SwiftUI New Onboarding Tour): WACO-54 Implement Brokerage carousel step (#2615)
At the beginning of the development of the new onboarding tour, a carousel had been added but only with placeholder views. In this PR the detail of the first 3 views is added.

Originally, this PR should only include the first view, but since the first 3 maintain the same structure, an independent development is not necessary for each of them, but they are built through a view builder.
It is important to note that the added images are those that were used in the figma design, and do not correspond to the final image to be used (which must still be defined by design).