feat(SwiftUI New Onboarding Tour): WACO-57 Implement Prices carousel step (#2671)
The price list consists of a header with texts and a pulsing circle (the pulse does not appear in the design, but I think it is a good addition) and a table with the price list.

The data in the table is currently mocked. A new table is expected to be added in the future with actual data. But at the moment this solution is presented until the definitive thing is ready to use.

The biggest difficulty with this view (and this is the reason why development took longer than expected) is that the table appears blurred at the bottom, but the page indicator is still present. The first solution I tried was to make the list transparent, so that the background gradient could be seen, however I couldn't achieve a transparent List in SwiftUI. The only thing that worked was to change the appearance of the UITableView and UITableViewCell, so that the List object was transparent. But using those proxies changes the appearance of all the tableviews of the app, and therefore for this solution to work well, more boilerplate code had to be added to cover that problem.

The second solution I tried was to put the list in a position deeper than the gradient view in a ZStack. This worked fine from a design point of view, but with the TabView being below the gradient view, the page indicator was no longer visible, so this solution was scrapped as well.

The next hack I tried was to put a transparent TabView on top of the original TabView, so that there was a visible page indicator while the original TabView was below the gradient view. But this solution added too much code to be able to coordinate both TabViews. Besides that in SwiftUI moving a tabView programmatically does not work well at all.

Finally the chosen solution was to change the list to a VStack within a ScrollView. In this way, it was possible to change the background to a transparent one and also add an opacity mask, in order to achieve the effect that is seen in the design.