feat(SwiftUI New Onboarding Tour): WACO-58 Add segmented button style (#2698)
Added the new component called "Minimal Double Button". This component corresponds to a segmented button that includes two buttons inside it. Added to the new Onboarding based on the designs.

The double button component did not exist at the time, so it has been added to UIComponents and can be reused. Its visual configuration, such as fonts and colors, is based on other elements of the same library, so that it is consistent with the other generic use buttons. The designs have been followed to comply with the three states of the button: enabled - disabled - pressed.