feat(SwiftUI New Onboarding Tour): WACO-66 Add Feature Flag and integration in Welcome Flow (#2721)
To add the new onboarding, the current implementation of the Welcome flow had to be analyzed. The new onboarding should replace the WelcomeView but at the same time be integrated with the rest of the screens that can appear as a product of the user tapping the buttons with the different alternatives.

The complexity of this operation is that the WelcomeView acts as a kind of base for the rest of the flow ... and the actions to show a sheet are based on changes of state. Therefore, it is not so easy to simply replace this view.

The first option I thought of was to create a completely new flow from the new onboarding, copying the rest of the screens so that the flow could be updated based on the new onboarding. I discarded this option because there is too much code and logic to copy ... and a bug could break all the things that are currently working.

The second option was to insert the new onboarding into the welcome view ... but this option complicated the use of flags and was also going to mess up the already saturated logic inside the WelcomeView.

The third option was to copy the logic of the WelcomeView into the new TourView. The problem with this solution is that the TourView was going to get dirty again in this case. It didn't seem like the correct development to me.

Finally I opted for the fourth option, which was to add an "adapter". This Adapter Mimics the functionality of the WelcomeView, without messing up the original WelcomeView, while using the TourView as the base view. In addition, I have taken advantage of improving the readability of the code so that it is cleaner compared to the original class.