feat(SwiftUI New Onboarding Tour): WACO-71 Implement Live Prices from Api (#2778)
Here basically the PriceRepositoryAPI has been used to obtain the different values ​​necessary in the presentation of the information. For example, the value of each crypto and the delta of the percentage of variation in the day. The architecture of the view has been modified to better model the new data compared to the previous simple mocks.