feat(Components): IOS-5535 ComponentLibrary Examples Setup (#2897)
It Adds an example app and configure the view that will contain all the examples from the library elements.
As the library is made up of multiple elements, a base view is created that will contain examples of each of these elements. Along with this, the base has been configured so that it is easy to add the examples.
Through this development, the structure is ready for the views to be added as they are developed.