feat(SwiftUI New Onboarding Tour): IOS-5647 & IOS-5643 Improve tour animations and performance (#2980)
In this PR several improvements and corrections are added to the new onboarding tour, mainly:

Header animation has been improved so that it increases / decreases its height linearly as the live prices view scrolls.

A transition is added in which the transparency of the list and the background gradient are blurred as the user scrolls.

The architecture of the Tour is modified / improved, in order to better comply with its operation.

It is also important to note that there was a performance problem when going from the second screen to the third in the main carousel of the tour. This was because at that moment the price list icons were loaded as images from the url resource. I tried different ways to solve this problem, but after several techniques the one that finally worked was to create the list in the first moment of the tour and add it to the stack of views of the tour, hidden by a white layer that serves as the background. In this way, when using this very same list later, the images were already loaded and the performance problem is eliminated. I know it's not a fancy solution, but given the limitations of Nuke and SwiftUI, I couldn't find a better way.