feat(AccountPicker): IOS-5706 Add Search to SwiftUI receive screen (#2979)
## Objective

Replicate search functionality from https://github.com/blockchain/wallet-ios-private/pull/2767 in `FeatureAccountPicker`

## Description

Added an optional search bar to the header.
Forwarded the search functionality from `AccountPickerInteractor` into the new account picker.

Refactoring of the state was necessary as searching something that already had a visible row would reset the balances states to `.loading`, but not call `.onAppear`. So instead of re-calling and reloading the balances, I've extracted the balances state out of the row itself and cached it on the main account picker state. This prevents the shimmering from re-appearing on already-loaded rows.

## How to Test

Enable `New SwiftUI Account Picker` feature flag, navigate to the "Receive" screen to view and use the search bar.

## Looms

https://www.loom.com/share/a5d0ab68a50e4f2fa65a2f0c4a67f910

## Merge Checklist

- [x] The PR uses a title supported by [.changelogrc](https://github.com/blockchain/My-Wallet-V3-iOS/blob/master/.changelogrc).
- [x] Code is formatted using `sh scripts/format-code.sh`.
- [x] UI components have accessibility identifiers and labels to support automated testing.
- [x] Areas of technical debt are marked with a `// TICKET:` comment that includes a ticket number.
- [x] You have added unit tests.
- [x] You have added sufficient documentation (descriptive comments).
- [x] New files added are within the correct directory. (e.g. if a file is required for unit tests to compile, be sure it is added to the tests target.)