feat(Components): IOS-5747 CircularIconButton (#3006)
## Objective

Button style used in navigation bars, and for close buttons.

## Description

Created `CircularIconButton`, which uses the same styling as `IconButton`, but the icon is inset in a filled circle.

## How to Test

ComponentLibraryExamples -> Buttons -> CircularIconButton

## Screenshot

See snapshots

## Notes

If this is merged before #2989 then that PR should be updated to change the examples + previews to use PrimaryNavigation. Same goes for this PR if that one is merged first.