feat(Components): IOS-5756 ComponentLibrary - refactor Pill Button Style (#3009)
In a first iteration, to achieve the design of the Pill Button style buttons (all these buttons look the same and only the color combination changes) a protocol was included (to avoid repeating code) in addition to the custom style. With this change, the protocol is eliminated and the entire button design is included within the style. I think that this is an improve in the code.