feat(Components): IOS-5518 ComponentLibrary - Double Primary Buttons (#3030)
This button can be understood as two pillbuttons together, in this case with the identical color combination as the PrimaryButton. It is necessary to add another custom ButtonStyle to achieve the visual effect that there are two buttons within one. Thanks to this, the "press" status can be seen for each button separately, also the rounded edges are for the entire button.