feat(Components): IOS-5728 ComponentLibrary - Balance Row (#3137)
The balance row is flexible in that it presents different versions of itself (based on the designs) depending on the parameters with which it is initialized. You can find four possibilities:

- default
- full (adds leading subtitle and graph view to the default version)
- full without graph
- full without leading subtitle

Not only are elements added / removed in the same positions, but depending on them the locations also change within the stacks. Due to this variability in the design, the code can look a bit complicated, but an attempt has been made to simplify the code as much as possible, complying with the design.