feat(Components): IOS-5532 ComponentLibrary - PageControl (#3246)
The PageControl corresponds to the set of circular indicators that indicate the selection status of an element with which it is linked. It is enough to indicate the set of elements to be controlled and the initial selection state. With this, the PageControl can handle the selection state or it indicates the selection change when it has been manipulated from the outside (manually changed the selection, for instance, with a tap on one of the elements.