feat(Components): IOS-5521 ComponentLibrary - PromoCard (#3399)
Add PromoCard to the Component Library.
The PromoCard has two different layouts: one when the user defines a control (button) and another when not. Not only the PrimaryButton inside the component appears, but also some spaces between views changes. The user can specified the title, message, control and the main icon.