chore(Tool): IOS-6292 Add Combine interoperability with async/await Task<Success, Failure> (#3606)
## Objective

Ahead of the guild discussions I am releasing this P/R to demonstrate how we can have interoperability with `Task`, the Swift Concurrency type allowing for async calls. 

As we move throughout the years Apple will be introducing more and more API with async - It is also a great way to simplify your asynchronous code. 

Please note, this is not a replacement for Combine but a tool to help it co-exist with the latest technology from Apple.

https://www.loom.com/share/440e7af2cea74cca8ff6ec3c49b56ddd

## Description

Add interop between Combine and Swift Concurrency 

Todo:
- [ ] ~~AsyncSequence~~
- [x] AsyncThrowingSequence

## Merge Checklist

- [x] The PR uses a title supported by [.changelogrc](https://github.com/blockchain/My-Wallet-V3-iOS/blob/master/.changelogrc).
- [x] Code is formatted using `sh scripts/format-code.sh`.