prototype(Namespace): IOS-6349 Initial Code and Schema Generation (#3687)
## Description

This is the first pull request for the implementation of [The blockchain.com semantic domain](https://www.notion.so/blockchaincom/The-blockchain-com-semantic-domain-f98ecc5c880a41c099dd345bb775fa79) and [Navigation and Deep Linking](https://www.notion.so/blockchaincom/Navigation-and-Deep-Linking-1e3f0bf56b8340128b177de661747454). 

Specifically, this includes the `blockchain.taskpaper` that [James Pettinger](mailto:jpettinger@blockchain.com) and myself have been working on to define user experiences which will be included in the redesign work.

The `gen` scheme can be run which will translate the taskpaper file into JSON and Swift code, eventually, I will also produce a generator for Kotlin and JavaScript.

To review this, please refer to `run.swift` as the bulk of this code/json is generated from the taskpaper. If you want to view the TaskPaper, please use any mind mapping tool (several exist on the Mac app store, as well as some on the web)

## Merge Checklist

- [x] The PR uses a title supported by [.changelogrc](https://github.com/blockchain/My-Wallet-V3-iOS/blob/master/.changelogrc).
- [x] Code is formatted using `sh scripts/format-code.sh`.