prototype(Namespace): IOS-6349 Add Tag.Reference (#3698)
## Description

To be able to reference indexed data we need to create a type which is able to store the underlying tag and the indices to what makes it unique. 

For example, we have `blockchain.user.email.address` where `blockchain.user` is a collection which can be indexed, if we were to look this data up inside of a database, or fetch something from an API we would need to know what user to lookup - this is where `indices` comes into play, indices are very much like the programming counterpart for indexing into collections.

A fully indexed id for the users email address would look like:
```
blockchain.user[7efa869d0364eea0cd0106a2ef4d1ae9eaec58fe62928c3f1af8fa8da9204ea0].email.address
```

We want to be able to pass context in various ways with differing levels of granularity, depending on the use-case including additional context is a great way to do that, since the key is long lasting and unique to the event. 

A more practical example might be when we are wanting to navigate to the transaction flow for BTC, we need to provide the asset information:

```
blockchain.ux.transaction.crypto.currency[BTC].buy
```

To review, look at Tag+Reference.swift

## Merge Checklist

- [x] The PR uses a title supported by [.changelogrc](https://github.com/blockchain/My-Wallet-V3-iOS/blob/master/.changelogrc).
- [x] Code is formatted using `sh scripts/format-code.sh`.