prototype(Namespace): IOS-6349 Add DeepLink DSL (#3701)
## Description

This piece of work introduces a small DSL for the URL which we can use in DEBUG when developing and testing deep links. The DSL will allow us to send events and set state. 

For example, we may have the following event to show the transaction flow:
```
blockchain.ux.deep_link.to.transaction.buy
```

We can trigger this by:
```
$ xcrun simctl openurl booted https://blockchain.page.link/app#blockchain.ux.deep_link.to.transaction.buy
```

On top of this, @crucheton-bc will be creating a rule processor to convert the URL's which Taro and team want into these events.

A rule to process the URL and translate it into one of our own system events could look like:

```
https://blockchain.com/app/transaction/buy?asset=BTC&amount=5000
```

```json
{
  "pattern": "^https://(www\\.)?blockchain.com/app/transaction/buy(/)?$",
  "event": "blockchain.ux.deep_link.to.transaction.buy",
  "parameters": {
    "asset": "blockchain.ux.deep_link.to.transaction.asset",
    "amount": "blockchain.ux.deep_link.to.transaction.amount"
  }
}
```

_note: these rules could live on Firebase Remote Config_

## Merge Checklist

- [x] The PR uses a title supported by [.changelogrc](https://github.com/blockchain/My-Wallet-V3-iOS/blob/master/.changelogrc).
- [x] Code is formatted using `sh scripts/format-code.sh`.