dev(Namespace): Always fetch shared values from the shared scope, always fetch user values from user scope (#4143)
An idea became apparent on how we should simplify access to preference values, anything marked as blockchain.session.state.shared.value should always be fetched from the shared scope and if the preference value is not marked with that, we always use the users scope - so we do not default back to the shared context if the value is missing. This is likely the expected behaviour for the consumer.