feat(Tx): IOS-6848 Fill the amount with the previous amount for that crypto asset, remember the preferred payment method for transactions (#4183)
## Objective

The average time for a customer to go into the brokerage flow and complete an order is over 1m30s - which is too slow. To help speed things up we want to try remove friction for the user.

1. Remember the users preferred payment method, and set it. This should always be the last payment method you chose.
2. Remember the previous transaction amount and pre-fill it on the enter amount screen.

## Description

Introduce two new ids for dealing with previous input amount:
```
blockchain.ux.transaction.source.target.previous.input.amount
blockchain.ux.transaction.source.target.previous.input.currency.code
```

`source` and `target` are both collections so we unique the amount on the pair which is being transacted. This value is stored in user defaults specific for the user, so their preference is remembered next time they open the app or log back in on the same device.

Remembering the payment method is consistent across any transaction and is specific to buy flow. This value can be set if we want to create a deep link into the transaction flow and pre-set the payment type to something like Apple Pay.
```
blockchain.ux.transaction.previous.payment.method.id
```

The input amount and payment method is committed to preferences when the customer taps "Preview Buy" and is shown the checkout. This is incase the transaction fails the customer can restore the exact same order as they previously made.