feat(Tx): Pre-fill amount on enter amount screen iff the previous transaction session for that asset failed (#4222)
## Description

If the previous transaction resulted in an error for that asset, we should restore the amount - this is cleared if the customer goes back to the checkout.

Also in this P/R:
1. dynamic subscription of references with unknown ids, to avoid race conditions when using state to manage ids - this publisher will always update references when the shared state receives any of the dynamic ids. Please see test.
2. Hot-path for references already resolved using the app

To test, please run the transaction flow - trigger an error and re-visit the flow. Your previous amount should be restored. During a normal dismiss or successful transaction, your amount will not be remembered. Your payment method will always be remembered.